Use html5lib due to bugs in internal HtmlParser
Python's html.parser.HtmlParser always decodes entityrefs, even when it should not. In particular, the link "index.php?option=database&params=2,1,0" was replaced with "index.php?option=database¶ms=2,1,0", because it interprets '&params' as '&para;ms' and replaces it with '¶ms'.
It is true that the error is on the cybenetics website: in a href, all ampersands should be encoded as &amp;. However, it is not the only website that has this error.
Sadly, I didn't find a way to override this behaviour in html.parser.HtmlParser. See https://stackoverflow.com/questions/79101170/
html5lib seems more reliable.