Option to download 115V, 230V or all PSUs
Introduce PREFERRED_WALL_POWER variable to select which PSU data should be downloaded, and which one is preferred if both are downloaded.
Note: even if you select to download 230V certified PSUs, the result may contain 115V tests. This is possible if a PSU is certified for both 115V and 230V, but the Light Load test is only done for one of the two voltages. This is allowed, because this is a supplementary test, not used for certification.